Update paper.md
In general a great inclusion of the wide range of problems that it can tackle!
Reading it I had difficulty to understand that the variable names are the attributes and we don't have a third variable name that we load. I thought of adding 'which we call'. What do you think @holgerteichgraeber ?